requirements.txt
lambda-git